requirements.txt
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
openpyxl>=3.1.2
beautifulsoup4>=4.12.2
click>=8.1.7
pyyaml>=6.0.1
lxml>=4.9.3
GitPython>=3.1.40
tqdm>=4.66.0
